Added a port of the original Observable "sheet" into javascript.
There is a bug in the port - that I can't see.

Differences between observables and conventional javascript are
getting in the way.

So I created a orignal subfolder which does things exactly things that the rust
port does but in javascript so I can compare itermediate values.

This is just a outline of the the javascript solution
with webpack config defined, and an initial sketch of the files.

Its not functional.